WIP: Sketch a material system
