--- src/graphics/texture.rs
@@ -1,75 +1,0 @@
-#[derive(Debug)]
-pub struct Texture {
-    pub(crate) texture: wgpu::Texture,
-    pub(crate) view: wgpu::TextureView,
-    pub(crate) sampler: wgpu::Sampler,
-}
-
-#[derive(Debug, Clone)]
-pub struct TextureData<'a> {
-    pub label: Option<String>,
-    pub pixels: &'a [u8],
-    pub format: wgpu::TextureFormat,
-    pub dimensions: (u32, u32),
-}
-
-impl<'a> TextureData<'a> {
-    pub fn upload(self, rend: &crate::Renderer) -> Texture {
-        let label = self.label.as_deref();
-        let size = wgpu::Extent3d {
-            width: self.dimensions.0,
-            height: self.dimensions.1,
-            depth_or_array_layers: 1,
-        };
-        let texture = rend.device.create_texture(&wgpu::TextureDescriptor {
-            label,
-            size,
-            mip_level_count: 1,
-            sample_count: 1,
-            dimension: wgpu::TextureDimension::D2,
-            format: self.format,
-            usage: wgpu::TextureUsages::TEXTURE_BINDING | wgpu::TextureUsages::COPY_DST,
-            view_formats: &[],
-        });
-
-        rend.queue.write_texture(
-            wgpu::ImageCopyTexture {
-                texture: &texture,
-                mip_level: 0,
-                origin: wgpu::Origin3d::ZERO,
-                aspect: wgpu::TextureAspect::All,
-            },
-            self.pixels,
-            wgpu::ImageDataLayout {
-                offset: 0,
-                bytes_per_row: Some(
-                    self.format
-                        .block_size(None)
-                        .expect("Incompatible texture format")
-                        * self.dimensions.0,
-                ),
-                rows_per_image: None,
-            },
-            size,
-        );
-
-        let view = texture.create_view(&wgpu::TextureViewDescriptor::default());
-
-        // TODO: get the sampler from gltf
-        let sampler = rend.device.create_sampler(&wgpu::SamplerDescriptor {
-            address_mode_u: wgpu::AddressMode::ClampToEdge,
-            address_mode_v: wgpu::AddressMode::ClampToEdge,
-            address_mode_w: wgpu::AddressMode::ClampToEdge,
-            mag_filter: wgpu::FilterMode::Linear,
-            min_filter: wgpu::FilterMode::Nearest,
-            mipmap_filter: wgpu::FilterMode::Nearest,
-            ..Default::default()
-        });
-
-        Texture {
-            texture,
-            view,
-            sampler,
-        }
-    }
-}
